change index[26],add login_register,and change in footer and header also[3]

+++ admin/ajax/features_facilities.php
@@ -43,9 +43,19 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_feature']];
 
-        $q = "DELETE FROM `features` WHERE `id`=?";
-        $res = delete($q,$values,'i');
-        echo $res;
+        $check_q = select('SELECT * FROM `room_features` WHERE `features_id`=?',[$frm_data['rem_feature']],'i');
+        if(mysqli_num_rows($check_q)==0)
+        {
+            $q = "DELETE FROM `features` WHERE `id`=?";
+            $res = delete($q,$values,'i');
+            echo $res;
+        }
+        else
+        {
+            echo "room_added";
+        }
+
+        
 
     }
 
@@ -108,20 +118,31 @@
         $frm_data = filteration($_POST);
         $values = [$frm_data['rem_facility']];
 
-        $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
-        $res = select($pre_q,$values,'i');
-        $img = mysqli_fetch_assoc($res);
-        
-        if(deleteImage($img['icon'],FACILITIES_FOLDER))
+        $check_q = select('SELECT * FROM `room_facilities` WHERE `facilities_id`=?',[$frm_data['rem_facility']],'i');
+
+        if(mysqli_num_rows($check_q)==0)
         {
-            $q = "DELETE FROM `facilities` WHERE `id`=?";
-            $res = delete($q,$values,'i');
-            echo $res;
+            $pre_q = "SELECT * FROM `facilities` WHERE `id`=?";
+            $res = select($pre_q,$values,'i');
+            $img = mysqli_fetch_assoc($res);
+            
+            if(deleteImage($img['icon'],FACILITIES_FOLDER))
+            {
+                $q = "DELETE FROM `facilities` WHERE `id`=?";
+                $res = delete($q,$values,'i');
+                echo $res;
+            }
+            else
+            {
+                echo 0;
+            }
         }
         else
         {
-            echo 0;
+            echo "room_added";
         }
+
+
